type cast for signal attributes (float64 -> int), implemented reference types for MIDI driver ports

+++ pkg/commands/commands.go
@@ -1,7 +1,7 @@
 package commands
 
 type TurnLightOnCommand struct {
-	KeyCode float64 `hubman:"key_code"`
+	KeyCode int `hubman:"key_code"`
 }
 
 func (c TurnLightOnCommand) Code() string {
@@ -13,7 +13,7 @@ func (c TurnLightOnCommand) Description() string {
 }
 
 type TurnLightOffCommand struct {
-	KeyCode float64 `hubman:"key_code"`
+	KeyCode int `hubman:"key_code"`
 }
 
 func (c TurnLightOffCommand) Code() string {
@@ -25,7 +25,7 @@ func (c TurnLightOffCommand) Description() string {
 }
 
 type SingleBlinkCommand struct {
-	KeyCode float64 `hubman:"key_code"`
+	KeyCode int `hubman:"key_code"`
 }
 
 func (c SingleBlinkCommand) Code() string {
@@ -37,7 +37,7 @@ func (c SingleBlinkCommand) Description() string {
 }
 
 type SingleReversedBlinkCommand struct {
-	KeyCode float64 `hubman:"key_code"`
+	KeyCode int `hubman:"key_code"`
 }
 
 func (c SingleReversedBlinkCommand) Code() string {
@@ -49,7 +49,7 @@ func (c SingleReversedBlinkCommand) Description() string {
 }
 
 type ContinuousBlinkCommand struct {
-	KeyCode float64 `hubman:"key_code"`
+	KeyCode int `hubman:"key_code"`
 }
 
 func (c ContinuousBlinkCommand) Code() string {
